+++ src/scenes/game/director.rb
