+++ src/scenes/game/player.rb
